refactor ipc functions
