+++ main/ipcEvents.ts
@@ -1,1 +1,7 @@
 export const authenticationString = 'authentication'
+
+export const getObs = 'get-obs'
+
+export const saveObs = 'save-obs'
+
+export const delObs = 'del-obs'
